pyproject.toml
[project]
name = "pyss"
version = "1.1.0"
authors = [
  { name="Nathan Fiscaletti", email="nate.fiscaletti@gmail.com" },
]
description = "Adds a utility for running pyss.yaml files."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies=[
    "pyyaml==6.0",
    "termcolor==2.2.0",
    "importlib-metadata==5.0.0",
    "packaging==21.3",
    "jsonschema==4.20.0",
]

[project.urls]
"Homepage" = "https://github.com/nathan-fiscaletti/pyss"

[project.scripts]
pyss = "pyss.main:main"